+++ packages/google_maps_flutter/google_maps_flutter_ios/ios/Classes/GoogleMapCircleController.m
@@ -60,43 +60,43 @@
 }
 
 - (void)interpretCircleOptions:(NSDictionary *)data {
-  NSNumber *consumeTapEvents = data[@"consumeTapEvents"];
-  if (consumeTapEvents && consumeTapEvents != (id)[NSNull null]) {
+  NSNumber *consumeTapEvents = FGMGetValueOrNilFromDict(data, @"consumeTapEvents");
+  if (consumeTapEvents) {
     [self setConsumeTapEvents:consumeTapEvents.boolValue];
   }
 
-  NSNumber *visible = data[@"visible"];
-  if (visible && visible != (id)[NSNull null]) {
+  NSNumber *visible = FGMGetValueOrNilFromDict(data, @"visible");
+  if (visible) {
     [self setVisible:[visible boolValue]];
   }
 
-  NSNumber *zIndex = data[@"zIndex"];
-  if (zIndex && zIndex != (id)[NSNull null]) {
+  NSNumber *zIndex = FGMGetValueOrNilFromDict(data, @"zIndex");
+  if (zIndex) {
     [self setZIndex:[zIndex intValue]];
   }
 
-  NSArray *center = data[@"center"];
-  if (center && center != (id)[NSNull null]) {
+  NSArray *center = FGMGetValueOrNilFromDict(data, @"center");
+  if (center) {
     [self setCenter:[FLTGoogleMapJSONConversions locationFromLatLong:center]];
   }
 
-  NSNumber *radius = data[@"radius"];
-  if (radius && radius != (id)[NSNull null]) {
+  NSNumber *radius = FGMGetValueOrNilFromDict(data, @"radius");
+  if (radius) {
     [self setRadius:[radius floatValue]];
   }
 
-  NSNumber *strokeColor = data[@"strokeColor"];
-  if (strokeColor && strokeColor != (id)[NSNull null]) {
+  NSNumber *strokeColor = FGMGetValueOrNilFromDict(data, @"strokeColor");
+  if (strokeColor) {
     [self setStrokeColor:[FLTGoogleMapJSONConversions colorFromRGBA:strokeColor]];
   }
 
-  NSNumber *strokeWidth = data[@"strokeWidth"];
-  if (strokeWidth && strokeWidth != (id)[NSNull null]) {
+  NSNumber *strokeWidth = FGMGetValueOrNilFromDict(data, @"strokeWidth");
+  if (strokeWidth) {
     [self setStrokeWidth:[strokeWidth intValue]];
   }
 
-  NSNumber *fillColor = data[@"fillColor"];
-  if (fillColor && fillColor != (id)[NSNull null]) {
+  NSNumber *fillColor = FGMGetValueOrNilFromDict(data, @"fillColor");
+  if (fillColor) {
     [self setFillColor:[FLTGoogleMapJSONConversions colorFromRGBA:fillColor]];
   }
 }